refact: Split App and PlayersList component
