+++ src/Component/PlayersList/index.js
@@ -1,0 +1,1 @@
+export { PlayersList } from './PlayersList'
